Gestion 360/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.1.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
